+++ include/ftl/task_counter.h
@@ -38,7 +38,7 @@ class TaskScheduler;
 class TaskCounter : public BaseCounter {
 
 public:
-	explicit TaskCounter(TaskScheduler *taskScheduler, unsigned const initialValue = 0, size_t const fiberSlots = NUM_WAITING_FIBER_SLOTS)
+	explicit TaskCounter(TaskScheduler *taskScheduler, unsigned const initialValue = 0, unsigned const fiberSlots = NUM_WAITING_FIBER_SLOTS)
 	        : BaseCounter(taskScheduler, initialValue, fiberSlots) {
 	}
